Move File type to top level types file
