+++ app/javascript/components/types.ts
@@ -1,0 +1,4 @@
+export type File = {
+  filename: string
+  content: string
+}
